requirements.txt
django>=2.2.1
channels>=2.2.0
channels-redis>=2.4.0
django-cors-headers>=3.0.2
djangorestframework>=3.9.4
djangorestframework-simplejwt>=4.3.0
djangochannelsrestframework>=0.0.3